+++ tests/system/src/test/java/test/com/sun/glass/ui/mac/MacOSSystemMenuMultiWindowFXOnlySwingFirst.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package test.javafx.stage;
+package test.com.sun.glass.ui.mac;
 
 import com.sun.javafx.tk.Toolkit;
 import javafx.application.Platform;
@@ -58,26 +58,23 @@ import javax.swing.JMenuItem;
 import javax.swing.SwingUtilities;
 import test.util.Util;
 
-public class MacOSSystemMenuMultiWindowTest extends MacOSSystemMenuTestBase {
+public class MacOSSystemMenuMultiWindowFXOnlySwingFirst extends MacOSSystemMenuTestBase {
 
     @Test
     public void test() throws InterruptedException, IOException {
-        initJavaFX(List.of(TEST_MENUS_0, TEST_MENUS_1, TEST_MENUS_2));
+        initSwing(List.of());
+        initJavaFX(List.of(TEST_MENUS_1, TEST_MENUS_3));
 
         focusJavaFX(0);
         List<Element> jfxElements = getMenusOfFocusedWindow();
-        compareMenus(jfxElements, TEST_MENUS_0);
+        compareMenus(jfxElements, TEST_MENUS_1);
 
         focusJavaFX(1);
         jfxElements = getMenusOfFocusedWindow();
-        compareMenus(jfxElements, TEST_MENUS_1);
-
-        focusJavaFX(2);
-        jfxElements = getMenusOfFocusedWindow();
-        compareMenus(jfxElements, TEST_MENUS_2);
+        compareMenus(jfxElements, TEST_MENUS_3);
 
         focusJavaFX(0);
         jfxElements = getMenusOfFocusedWindow();
-        compareMenus(jfxElements, TEST_MENUS_0);
+        compareMenus(jfxElements, TEST_MENUS_1);
     }
 }
