+++ tests/system/src/test/java/test/com/sun/glass/ui/mac/MacOSSystemMenuMultiWindowFXOnlySwingLast.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package test.javafx.stage;
+package test.com.sun.glass.ui.mac;
 
 import com.sun.javafx.tk.Toolkit;
 import javafx.application.Platform;
