+++ tests/system/src/test/java/test/com/sun/glass/ui/mac/MacOSSystemMenuMultiWindowWithSwingFirstTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, 2024, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package test.javafx.stage;
+package test.com.sun.glass.ui.mac;
 
 import com.sun.javafx.tk.Toolkit;
 import javafx.application.Platform;
@@ -58,53 +58,31 @@ import javax.swing.JMenuItem;
 import javax.swing.SwingUtilities;
 import test.util.Util;
 
-
-public class MacOSSystemMenuSetWithSwingTestFirst extends MacOSSystemMenuTestBase {
-
-    private CountDownLatch latch;
+public class MacOSSystemMenuMultiWindowWithSwingFirstTest extends MacOSSystemMenuTestBase {
 
     @Test
     public void test() throws InterruptedException, IOException {
-        initSwing(List.of(TEST_MENUS_0));
-        initJavaFX(List.of(TEST_MENUS_1));
+        initSwing(List.of(TEST_MENUS_0, TEST_MENUS_2));
+        initJavaFX(List.of(TEST_MENUS_1, TEST_MENUS_3));
 
         focusJavaFX(0);
-        runOnFXThread(() -> javaFXMenuBars.get(0).setUseSystemMenuBar(false));
+        List<Element> jfxElements = getMenusOfFocusedWindow();
+        compareMenus(jfxElements, TEST_MENUS_1);
 
-        List<Element> elements = getMenusOfFocusedWindow();
-        compareMenus(elements, List.of());
-
-        runOnFXThread(() -> javaFXMenuBars.get(0).setUseSystemMenuBar(true));
-        elements = getMenusOfFocusedWindow();
-        compareMenus(elements, TEST_MENUS_1);
+        focusJavaFX(1);
+        jfxElements = getMenusOfFocusedWindow();
+        compareMenus(jfxElements, TEST_MENUS_3);
 
         focusSwing(0);
-        runOnFXThread(() -> javaFXMenuBars.get(0).setUseSystemMenuBar(false));
-        elements = getMenusOfFocusedWindow();
-        compareMenus(elements, TEST_MENUS_0);
+        List<Element> swingElements = getMenusOfFocusedWindow();
+        compareMenus(swingElements, TEST_MENUS_0);
 
-        focusJavaFX(0);
-        elements = getMenusOfFocusedWindow();
-        compareMenus(elements, List.of());
+        focusJavaFX(1);
+        jfxElements = getMenusOfFocusedWindow();
+        compareMenus(jfxElements, TEST_MENUS_3);
 
-        focusSwing(0);
-        runOnFXThread(() -> javaFXMenuBars.get(0).setUseSystemMenuBar(true));
-        elements = getMenusOfFocusedWindow();
-        compareMenus(elements, TEST_MENUS_0);
-
-        focusJavaFX(0);
-        elements = getMenusOfFocusedWindow();
-        compareMenus(elements, TEST_MENUS_1);
-    }
-
-    private void runOnFXThread(Runnable runnable) throws InterruptedException {
-        latch = new CountDownLatch(1);
-
-        Platform.runLater(() -> {
-            runnable.run();
-            latch.countDown();
-        });
-
-        latch.await();
+        focusSwing(1);
+        swingElements = getMenusOfFocusedWindow();
+        compareMenus(swingElements, TEST_MENUS_2);
     }
 }
