+++ tests/system/src/test/java/test/com/sun/glass/ui/mac/MacOSSystemMenuTestBase.java
@@ -1,4 +1,28 @@
-package test.javafx.stage;
+/*
+ * Copyright (c) 2025, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+package test.com.sun.glass.ui.mac;
 
 import com.sun.javafx.tk.Toolkit;
 import javafx.application.Platform;
@@ -29,7 +53,6 @@ import javax.swing.JMenuBar;
 import javax.swing.JMenuItem;
 import javax.swing.SwingUtilities;
 
-import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -139,11 +162,11 @@ public class MacOSSystemMenuTestBase {
 
     private CountDownLatch latch = null;
 
-    /////////////////////////////////////////////////////////
-    ///
-    /// Helpers for creation and focusing of windows
-    ///
-    /////////////////////////////////////////////////////////
+    /***************************************************
+     *
+     * Helpers for creation and focusing of windows
+     *
+     **************************************************/
 
     protected void initJavaFX(List<List<Element>> menus) {
         initJavaFX(false, menus);
@@ -278,11 +301,11 @@ public class MacOSSystemMenuTestBase {
         return menu;
     }
 
-    /////////////////////////////////////////////////////////
-    ///
-    /// Helpers for system menu comparison
-    ///
-    /////////////////////////////////////////////////////////
+    /***************************************************
+     *
+     * Helpers for system menu comparison
+     *
+     **************************************************/
 
     /**
      * Compares the app menus of the provided menu bars. The app menu
@@ -399,11 +422,11 @@ public class MacOSSystemMenuTestBase {
         }
     }
 
-    /////////////////////////////////////////////////////////
-    ///
-    /// Helpers for synchronization
-    ///
-    /////////////////////////////////////////////////////////
+    /***************************************************
+     *
+     * Helpers for synchronization
+     *
+     **************************************************/
 
     private void initLock() {
         latch = new CountDownLatch(1);
@@ -417,11 +440,5 @@ public class MacOSSystemMenuTestBase {
 
     private void releaseLock() {
         latch.countDown();
-    }
-
-    protected void sleep(long millis) {
-        try {
-            Thread.sleep(millis);
-        } catch (Exception e) {}
     }
 }
